改了一些采样的bug将pwm输出改为普通的添加了电流环频率设置

+++ XDr/Code/XDrive/FOC/Config/system_parameters.h
@@ -16,8 +16,8 @@
 #define SYSTEM_DESC_str \
     Description "," VERSION "," AUTHOR "," Frequency_string "," MAX_CURRENT_string "," MAX_Voltage_string "," MIN_Voltage_string "," MAX_Temperature_string
 
-#define fpwm 20000       // 20kHz
-#define Tpwm 1.0f / fpwm // 50us
+#define fpwm (unsigned int)20000 // 20kHz
+#define Tpwm 1.0f / fpwm         // 50us
 #define ticpwm 2099
 #define Tcon Tpwm
